feat: My account, account page, extract model and fix build, repackage
